+++ src/stress.cpp
@@ -76,27 +76,27 @@ void dft_force::operator-=(const dft_force &st) {
   if (diag && st.diag) *diag -= *st.diag;
 }
 
-static void stress_sum(int Nfreq, double *F, const dft_chunk *F1, const dft_chunk *F2) {
+static void stress_sum(size_t Nfreq, double *F, const dft_chunk *F1, const dft_chunk *F2) {
   for (const dft_chunk *curF1 = F1, *curF2 = F2; curF1 && curF2;
        curF1 = curF1->next_in_dft, curF2 = curF2->next_in_dft) {
     complex<realnum> extra_weight(real(curF1->extra_weight), imag(curF1->extra_weight));
     for (size_t k = 0; k < curF1->N; ++k)
-      for (int i = 0; i < Nfreq; ++i)
+      for (size_t i = 0; i < Nfreq; ++i)
         F[i] += real(extra_weight * curF1->dft[k * Nfreq + i] * conj(curF2->dft[k * Nfreq + i]));
   }
 }
 
 double *dft_force::force() {
-  const int Nfreq = freq.size();
+  const size_t Nfreq = freq.size();
   double *F = new double[Nfreq];
-  for (int i = 0; i < Nfreq; ++i)
+  for (size_t i = 0; i < Nfreq; ++i)
     F[i] = 0;
 
   stress_sum(Nfreq, F, offdiag1, offdiag2);
   stress_sum(Nfreq, F, diag, diag);
 
   double *Fsum = new double[Nfreq];
-  sum_to_all(F, Fsum, Nfreq);
+  sum_to_all(F, Fsum, int(Nfreq));
   delete[] F;
   return Fsum;
 }
